Move CardPanel to top level components
