+++ src/scenes/SignUp/SignUpFormWizard/index.js
@@ -2,9 +2,10 @@ import React from 'react';
 // import BlogStep from './Blog';
 // import StartStep from './Start';
 // import NicknamesStep from './Nicknames';
+import CardPanel from 'components/CardPanel';
 
 const Wizard = () => {
-  return <div>SIGN UP!!!</div>;
+  return <CardPanel colSizes="s12">SIGN UP!!!</CardPanel>;
 };
 
 export default Wizard;
